+++ packages/apptive_grid_heinzelmen/test/src/attachment_image_test.dart
@@ -173,7 +173,7 @@ class _MockHttpClient extends Mock implements HttpClient {
   }
 
   @override
-  set autoUncompress(bool _autoUncompress) {}
+  set autoUncompress(bool autoUncompress) {}
 }
 
 class _MockHttpClientRequest extends Mock implements HttpClientRequest {}