+++ packages/apptive_grid_heinzelmen/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.3
+* Fixed the Stage Banner to work with the generics from the ConfigurationChangeNotifier
+* Update to flutter_lints 2 and fix warnings
+
+## 0.0.3-alpha.1 - Fixed Stage Banner
+* Fixed the Stage Banner to work with the generics from the ConfigurationChangeNotifier
+* Upgraded to new ApptiveGrid package version
+
 ## 0.0.2 - Fixed Environment Switch
 * Fixed the Environment Switch to work with the generics from the ConfigurationChangeNotifier